apps/http-server/plugins/protocol_tof: refactor protocol_tof.c and added modeinfo
